requirements.txt
fastapi[standard]==0.115.0
pydantic-settings==2.5.2
beanie==1.26.0
passlib==1.7.4
python-jose==3.3.0
pytest==7.4.4
pytest-asyncio==0.24.0
httpx==0.26.0
